% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beds_state}
\alias{beds_state}
\title{Hospital beds annual series}
\format{A data frame with six variables}
\source{
[http://kff.org/other/state-indicator/beds-by-ownership]()
}
\usage{
beds_state
}
\description{
state-year level hospital beds per 1,000 Population.
}
\keyword{datasets}

