% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitals-package.R
\docType{package}
\name{hospitals-package}
\alias{hospitals-package}
\alias{_PACKAGE}
\title{hospitals: Portuguese NHS hospitals}
\description{
A dataset of the Portuguese NHS hospitals.
}
\author{
\strong{Maintainer}: Ramiro Magno \email{ramiro.magno@gmail.com} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Faculty of Economics, University of Algarve [copyright holder, funder]
}

}
\keyword{internal}
